requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
celery==5.3.4
redis==5.0.1
python-telegram-bot==20.7
python-dotenv==1.0.0
django-filter==23.3
drf-spectacular==0.26.5
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
flake8==6.1.0
